requirements.txt
openai==1.74.0
fpdf2==2.7.6
requests
beautifulsoup4
python-dotenv